Add files via upload
Code that currently allows a user to drag a object and prints the (x,y) cords of where the box ends up. There is also currently a 5x5 grid. Currently there is a small red box in the middle of the two "ships". I plan on making this in to as many dots for as many 'blocks' the ship takes up, (I.E,  A ships is 1 'grid' wide and 3 'grids' tall, it will have 3 dots one for every square it has). This way when I add these dots to the middle of every spot on the grid when a person "lets go" of a "ship" it will 'snap' into place on the grid allowing me to pull that data and build a background 2D array which will be the starting game board.